requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
boto3==1.34.0
cryptography==42.0.0
python-dotenv==1.0.0
werkzeug==3.0.1
bcrypt==4.1.2
requests==2.31.0
